moved consume into module that includes exponential backoff and caching
