+++ requirements.txt
@@ -1,4 +1,5 @@
 autopep8==1.2.4
+backoff==1.4.3
 boto3==1.4.4
 coverage==4.1
 Django==1.11
